requirements.txt
python-dotenv
openai
python-bithumb
pyupbit